+++ net.sourceforge.pmd.eclipse.plugin.test/src/main/java/net/sourceforge/pmd/eclipse/runtime/properties/ProjectPropertiesModelTest.java
@@ -15,12 +15,15 @@ import java.net.URL;
 import java.net.URLClassLoader;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
+import java.util.regex.Pattern;
 
 import org.apache.commons.io.IOUtils;
 import org.eclipse.core.resources.IFile;
@@ -523,9 +526,11 @@ public class ProjectPropertiesModelTest {
      * </ul>
      *
      * @throws Exception
+     * @deprecated Since 7.21.0. Tests the deprecated {@link IProjectProperties#getAuxClasspath()} method.
      */
     @Test
-    public void testProjectClasspath() throws Exception {
+    @Deprecated
+    public void testProjectClasspathClassloader() throws Exception {
         IProject otherProject = EclipseUtils.createJavaProject("OtherProject");
         additionalProjects.add(otherProject);
         IFile sampleLib1 = otherProject.getFile("sample-lib1.jar");
@@ -606,5 +611,96 @@ public class ProjectPropertiesModelTest {
 
         // no remaining urls
         Assert.assertTrue(urls.isEmpty());
+    }
+
+    /**
+     * Project structure:
+     * <ul>
+     * <li>this.testProject "ProjectPropertiesModelTest": main project, with build path, contains lib/sample-lib3.jar</li>
+     * <li>otherProject "OtherProject": contains sample-lib1.jar, sample-lib2.jar</li>
+     * <li>otherProject2 "OtherProject2": ProjectPropertiesModelTest depends on this</li>
+     * <li>externalProject "ExternalProject": not stored within workspace, contains sample-lib4.jar</li>
+     * </ul>
+     *
+     * @throws Exception
+     */
+    @Test
+    public void testProjectClasspath() throws Exception {
+        IProject otherProject = EclipseUtils.createJavaProject("OtherProject");
+        additionalProjects.add(otherProject);
+        IFile sampleLib1 = otherProject.getFile("sample-lib1.jar");
+        sampleLib1.create(IOUtils.toInputStream("", "UTF-8"), false, null);
+        File realSampleLib1 = sampleLib1.getLocation().toFile().getCanonicalFile();
+        IFile sampleLib2 = otherProject.getFile("sample-lib2.jar");
+        sampleLib2.create(IOUtils.toInputStream("", "UTF-8"), false, null);
+        File realSampleLib2 = sampleLib2.getLocation().toFile().getCanonicalFile();
+
+        IFolder libFolder = this.testProject.getFolder("lib");
+        libFolder.create(false, true, null);
+        IFile sampleLib3 = libFolder.getFile("sample-lib3.jar");
+        sampleLib3.create(IOUtils.toInputStream("", "UTF-8"), false, null);
+        File realSampleLib3 = sampleLib3.getLocation().toFile().getCanonicalFile();
+
+        IProject otherProject2 = EclipseUtils.createJavaProject("OtherProject2");
+        additionalProjects.add(otherProject2);
+        // build the project, so that the output folder "bin/" is created
+        otherProject2.build(IncrementalProjectBuilder.FULL_BUILD, null);
+
+        IProject externalProject = ResourcesPlugin.getWorkspace().getRoot().getProject("ExternalProject");
+        additionalProjects.add(externalProject);
+        Assert.assertFalse("Project must not exist yet", externalProject.exists());
+        java.nio.file.Path externalProjectDir = Files.createTempDirectory("pmd-eclipse-plugin");
+        IProjectDescription description = externalProject.getWorkspace().newProjectDescription("ExternalProject");
+        description.setLocation(Path.fromOSString(externalProjectDir.toString()));
+        externalProject.create(description, null);
+        externalProject.open(null);
+        IFile sampleLib4 = externalProject.getFile("sample-lib4.jar");
+        sampleLib4.create(IOUtils.toInputStream("", "UTF-8"), false, null);
+        File realSampleLib4 = sampleLib4.getLocation().toFile().getCanonicalFile();
+
+        // build the project, so that the output folder "bin/" is created
+        this.testProject.build(IncrementalProjectBuilder.FULL_BUILD, null);
+
+        IFile file = this.testProject.getFile(".classpath");
+        String newClasspathContent = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
+            + "<classpath>\n"
+            + "    <classpathentry kind=\"src\" path=\"src\"/>\n"
+            + "    <!-- <classpathentry kind=\"con\" path=\"org.eclipse.jdt.launching.JRE_CONTAINER\"/> -->\n"
+            + "    <classpathentry combineaccessrules=\"false\" kind=\"src\" path=\"/OtherProject2\"/>\n"
+            + "    <classpathentry kind=\"lib\" path=\"/OtherProject/sample-lib1.jar\"/>\n"
+            + "    <classpathentry kind=\"lib\" path=\"" + realSampleLib2.getAbsolutePath() + "\"/>\n"
+            + "    <classpathentry kind=\"lib\" path=\"lib/sample-lib3.jar\"/>\n"
+            + "    <classpathentry kind=\"output\" path=\"bin\"/>\n"
+            + "    <classpathentry kind=\"lib\" path=\"/ExternalProject/sample-lib4.jar\"/>\n"
+            + "</classpath>\n";
+        file.setContents(IOUtils.toInputStream(newClasspathContent, "UTF-8"), 0, null);
+        // refresh, so that changed .classpath file is considered
+        this.testProject.refreshLocal(IResource.DEPTH_INFINITE, null);
+        // rebuild again, so that changed classpath is configured on java project
+        this.testProject.build(IncrementalProjectBuilder.FULL_BUILD, null);
+
+        final IProjectPropertiesManager mgr = PMDPlugin.getDefault().getPropertiesManager();
+        IProjectProperties model = mgr.loadProjectProperties(this.testProject);
+        List<String> classpath = new ArrayList<>(Arrays.asList(model.getClasspath().split(Pattern.quote(File.pathSeparator))));
+
+        Assert.assertEquals("Found these paths: " + classpath, 6, classpath.size());
+
+        // own project's output folder
+        Assert.assertTrue(classpath.remove(
+                new File(this.testProject.getLocation().toFile().getAbsoluteFile(), "bin").toString()));
+        // output folder of other project 2 (project dependency)
+        Assert.assertTrue(classpath.remove(
+                new File(otherProject2.getLocation().toFile().getAbsoluteFile(), "bin").toString()));
+        // sample-lib1.jar stored in OtherProject
+        Assert.assertTrue(classpath.remove(realSampleLib1.toString()));
+        // sample-lib2.jar referenced with absolute path
+        Assert.assertTrue(classpath.remove(realSampleLib2.toString()));
+        // sample-lib3.jar stored in own project folder lib
+        Assert.assertTrue(classpath.remove(realSampleLib3.toString()));
+        // sample-lib4.jar stored in external project folder outside of workspace
+        Assert.assertTrue(classpath.remove(realSampleLib4.toString()));
+
+        // no remaining urls
+        Assert.assertTrue(classpath.isEmpty());
     }
 }